kv-cache: fix narrowing conversion breaking strict-narrowing compilers
llama_kv_cache_context::get_kv_stream_active_targets() built
{ kv, n_kv } where n_kv is int32_t but llama_kv_stream_active_target's
second field is uint32_t - a silent narrowing conversion GCC only warns
about but clang-cl (Windows CI) treats as a hard error
(-Wc++11-narrowing), failing every windows build variant. Explicit cast.

+++ src/llama-kv-cache.cpp
@@ -3368,7 +3368,7 @@ std::vector<llama_kv_stream_active_target> llama_kv_cache_context::get_kv_stream
     if (kv->get_kv_stream_targets().empty()) {
         return {};
     }
-    return { { kv, n_kv } };
+    return { { kv, (uint32_t) n_kv } };
 }
 
 ggml_type llama_kv_cache_context::type_k() const {
